feat: BarChart component
